pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "evoagents"
version = "0.1.0"
description = "Agents that evolve their own skills. Self-healing multi-agent systems."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "EvoAgents Contributors" }]
keywords = ["agents", "llm", "self-healing", "prompt-engineering", "multi-agent", "evoagents"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "openai>=1.70.0",
    "anthropic>=0.20.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
evoagents = "evoagents.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["evoagents"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
